+++ cfgov/core/jinja2tags.py
@@ -2,6 +2,7 @@ from flags.template_functions import flag_disabled, flag_enabled
 from jinja2.ext import Extension
 
 from core.templatetags.svg_icon import svg_icon
+from core.utils import signed_redirect, unsigned_redirect
 
 
 class CoreExtension(Extension):
@@ -10,6 +11,9 @@ class CoreExtension(Extension):
         self.environment.globals.update({
             'flag_enabled': flag_enabled,
             'flag_disabled': flag_disabled,
+
+            'signed_redirect': signed_redirect,
+            'unsigned_redirect': unsigned_redirect,
 
             'svg_icon': svg_icon,
         })
