+++ cfgov/core/tests/test_jinja2tags.py
@@ -6,28 +6,14 @@ from django.test import TestCase, override_settings
 from core.tests.templatetags.test_svg_icon import VALID_SVG
 
 
-_TEST_TEMPLATES = [
-    {
-        'NAME': 'test',
-        'BACKEND': 'django.template.backends.jinja2.Jinja2',
-        'OPTIONS': {
-            'extensions': [
-                'core.jinja2tags.filters',
-            ],
-        },
-    },
-]
-
-
 @override_settings(
-    TEMPLATES=_TEST_TEMPLATES,
     STATICFILES_DIRS=[
         os.path.join(os.path.dirname(__file__), 'staticfiles'),
     ]
 )
 class SvgIconTests(TestCase):
     def setUp(self):
-        self.jinja_engine = engines['test']
+        self.jinja_engine = engines['wagtail-env']
 
     def test_jinja_tag(self):
         template = self.jinja_engine.from_string('{{ svg_icon("test") }}')
@@ -39,13 +25,10 @@ class SvgIconTests(TestCase):
             template.render()
 
 
-@override_settings(
-    TEMPLATES=_TEST_TEMPLATES,
-    FLAGS={'MY_FLAG': {'boolean': 'True'}}
-)
+@override_settings(FLAGS={'MY_FLAG': {'boolean': 'True'}})
 class FeatureFlagTests(TestCase):
     def setUp(self):
-        self.jinja_engine = engines['test']
+        self.jinja_engine = engines['wagtail-env']
 
     def test_flag_enabled_tag(self):
         template = self.jinja_engine.from_string(
