+++ cfgov/v1/tests/jinja2tags/test_jinja2tags.py
@@ -6,7 +6,7 @@ from django.test import RequestFactory, TestCase, override_settings
 from model_mommy import mommy
 
 from v1.atomic_elements.atoms import ImageBasic
-from v1.jinja2tags import date_formatter, email_popup, image_alt_value
+from v1.jinja2tags import email_popup, image_alt_value
 from v1.models import CFGOVImage, CFGOVRendition
 
 
@@ -54,41 +54,16 @@ class TestImageAltValue(TestCase):
         self.assertEqual(image_alt_value(value), 'Alt text on block')
 
 
-class TestDateFormatter(TestCase):
-    def test_text_format_date(self):
-        test_date = date(2018, 9, 5)
-        output = date_formatter(test_date, text_format=True)
-        self.assertIn('Sept. 5, 2018', output)
-
-    def test_default_format_date(self):
-        test_date = date(2018, 9, 5)
-        output = date_formatter(test_date, text_format=False)
-        self.assertIn('Sep 05, 2018', output)
-
-
 class TestEmailPopup(TestCase):
     def test_email_popup_defined_and_returns_empty_for_no_popup(self):
         request = RequestFactory().get('/page/without/a/popup')
         self.assertEqual(email_popup(request), '')
 
 
-@override_settings(
-    TEMPLATES=[
-        {
-            'NAME': 'test',
-            'BACKEND': 'django.template.backends.jinja2.Jinja2',
-            'OPTIONS': {
-                'extensions': [
-                    'v1.jinja2tags.filters',
-                ],
-            },
-        },
-    ]
-)
 class TestIsFilterSelected(TestCase):
     def setUp(self):
         self.factory = RequestFactory()
-        self.jinja_engine = engines['test']
+        self.jinja_engine = engines['wagtail-env']
 
     def _render_template_with_request(self, request):
         s = '{{ is_filter_selected("foo", "bar") }}'
