+++ cfgov/v1/tests/templates/foo/bar/test.html
@@ -1,0 +1,4 @@
+include "include.html": {% include "include.html" %}
+include "bar/include.html": {% include "bar/include.html" %}
+include "foo/bar/include.html": {% include "foo/bar/include.html" %}
+include "/include.html": {% include "/include.html" %}
