+++ cfgov/v1/tests/templates/foo/include.html
@@ -1,0 +1,1 @@
+include from foo
